Org: Introduce a CANP state
CANP = canceling just because I'm canceling the parent (so if I restore
the parent, I'll want to restore this).

Here are some notes on the thinking that inspired this:

Problem: I want to cancel a task that has subtasks, but not go through
and cancel all of the sub-tasks – in part in case I want to restore this
later: I want it to be clear that these subtasks are not themselves
still canceled if I restore the parent task.

Normally, to cancel or mark a task as complete, you need to cancel or
mark as complete all subtasks (with the settings I have, anyway). You
can force marking the parent as complete with C-u C-u C-u C-c C-t. But
then the statistics in the parent project will think there are still
some incomplete tasks. I thought about changing the way I track those
statistics to only count direct children, but that then excludes
archived things.

Maybe the best solution is to introduce a new state, like canceled, but
for this specific situation.