Add imenu to menubar
I had removed this before because I noticed it was causing some typing
lag. However, I'm feeling like it would be helpful, and from some
further experimentation it seems like the typing lag may only be an
issue if imenu-auto-rescan is set to t. So I'm going to try having imenu
in the menubar but without imenu-auto-rescan, and see if there are major
downsides to this configuration.

I'm actually not sure how helpful the imenu menubar thing will be, but
I'm thinking it might make sense to set imenu-auto-rescan to nil anyway,
in case it is responsible for some lag even without imenu in the
menubar.