Revert "Improved kill-sentence behavior"
This reverts commit 4fac153f607f366b92664b457f12fcb5af41aaf2.

I was finding that there are pros and cons of this change. One con, for
example, is that it doesn't work right when deleting the next sentence
if the cursor is before the leading space before that sentence.

So, since this isn't definitively better, I want to try more with the
out-of-the-box behavior of kill-sentence.