Put schedule at top of org agenda views
The main motivation for this was to reorganize the new "w" view, where I
want the SOON/NEXT/PROG items above the projects. But I realized it
could be good to have scheduled items at the top in general, to make
them more visible (if I ever start to use them).