Add an org speed command to insert a new todo
The original motivation was wanting to make it easier to add a task
after the parent of the current task. I thought about making a dedicated
command for that, but then thought it would be easiest to do that in two
steps: go to the parent, then insert a todo. Now this can be done with
C-c C-u (existing command to go to parent) then space. The ability to
add a task with space seems very nice in general (similar to what I can
do in Things).

Also change the org speed command to display outline path from " " to
"o" so I can use " " for my new command to insert a todo.