Set org-cycle-include-plain-lists to nil
Motivation: I want to be able to collapse a heading even if I'm inside a
list; this can currently be done with C-c Tab, but being able to do it
with just Tab is handy. I don't feel like I will often want to collapse
list items; if I do, I could make a function to temporarily let-bind
org-cycle-include-plain-lists before calling org-cycle and bind that to
a key.