Org: set org-inlinetask-show-first-star to t
The main motivation is to help delineate the body of inline tasks, since
the body isn't indented (which may be a bug? but that's the behavior in
out-of-the-box org with uncustomized emacs as well as with my
customizations). The fact that it makes inline tasks stand out more
seems like a possibly good thing as well.