Fix problem with magit hanging sometimes when viewing a commit
I don't love these workarounds (explicitly listing modes for which we
enable which-function, and no longer automatically adding imenu to the
menubar), but they are better than having emacs freeze sometimes. And
actually, maybe I'll find it's not such a bad thing to only have
which-function-mode enabled in select modes: that might speed things up
in other modes.