org-agenda: show extra component in outline path for archived items
I want to show the area and the project in the outline path in the
header. For unarchived items, this can be done by showing the first two
components of the outline path, but for archived items the path is
AREA/Archive/PROJECT - so I need to show a 3rd path component.