Use company-manual-begin instead of company-complete to invoke company
With company-complete, some hard-to-predict amount of text would get
filled in, which I think may slow me down rather than speeding me up.

Another nice advantage of using company-manual-begin is that I can use
it to confirm that something I just typed is a valid symbol: with
company-manual-begin, if it's valid, I'll see a clear indication of
that; with company-complete, nothing happens if point is at the end of a
complete symbol.