Use just 4 outline colors that repeat
I was finding that some of the later ones looked too similar to each
other and to the earlier ones. It seems easier to come up with 4 colors
that look distinct and then have these repeat -- figuring that I'm not
going to get too confused about whether a header is level 1, 5 or 9 (or
2, 6 or 10, etc.) - but I do get confused about whether a header
is (e.g.) level 6 or 8.

Using green (005000) would allow more distinct colors (e.g., by getting
rid of purple4, which is somewhat similar to blue4) or getting rid of
the not very attractive sienna4, but I don't really like the look of
green.