outline-2 face: use green instead of sienna
I don't love the look of the green, but it's easier to read than the
sienna and avoids confusion with outline-4.

(I may decide to now swap 1/2 and 3/4; I'm not sure. The main point of
that would be to avoid color confusion until I get to 6. Or maybe I can
come up with a totally different color from the current purple4 to avoid
confusion with blue4.)