Company: also use TAB to complete selection
Motivation: my thumbs have been hurting a bit lately, and I wonder if
one contributing factor is that I sometimes hold down M or A-M for a
second in order to do a double application of M-SPC or A-M-SPC,
particularly now that I need two applications to do the completion when
there is just one completion option (since now I use
company-manual-begin instead of company-complete). I'm hoping that, by
changing my typical usage to use TAB instead of M-SPC or A-M-SPC to do
the final completion, this won't be an issue.