Improve vdiff scrolling
(1) It doesn't work to scroll a window by small amounts if there are
virtual removed lines at the top of the window. So if we try and fail to
scroll a given vdiff window, switch to the other window and try the
scrolling there.

(2) vdiff's synchronized scrolling doesn't work when you try to do mouse
scrolling of one window when the other window is active. So force mouse
scrolling to only work on the active window in vdiff mode. (See also
https://github.com/justbur/emacs-vdiff/issues/29.) This is especially
important to make mouse scrolling usable given that we're doing (1),
since which window is active can now change frequently while scrolling.