Make outline mode ellipses stand out much more
Customize the look of ellipses used in outline mode: I want them to
stand out a lot more than they do by default. The motivation is my use
of outline-minor-mode in programming modes: I want to make it obvious
when some code is hidden. (Changing the face might be better, but I
can't figure out how to do that.)