Don't turn outline minor mode on automatically; add a keybinding for it
My new plan is to generally keep outline minor mode off, and just turn
it on as needed. That way, I can see from the mode line if there is a
possibility that things are hidden due to outline-minor-mode (and
conversely, if I don't see Outl in the mode line, then I can be
confident that nothing is hidden due to it).