Only run org-latex-impatient previewing manually
This started from being worried that the constant checking of whether
we're in a latex region would have a performance impact. I doubt that's
a major issue. But then from some playing around, I wasn't loving the
auto-re-rendering, because sometimes it would trigger when the equation
was in error. So I like the idea of just triggering this manually.