Turn off lsp-enable-indentation
In C++, this led to indent-region having different rules from when I'm
typing and hitting tab, and I don't like that these two things operate
differently. I haven't experimented with other languages, but I think
it's best to use the same rules for indenting regions as for when I'm
actually typing.