Org: Change speed commands for archiving
Make 'a' archive to archive sibling (matching what is now C-c a) and 'A'
toggle archive tag (matching what is now C-c A). ('a' was originally
org-archive-subtree-default-with-confirmation. There may be value in
having a speed command do whatever the default org archive is here -
with or without confirmation - but I'll consider coming back to that
idea later, once I have figured out how to set up default archiving the
way I want it.)