Disable the M-S-return binding for my-org-insert-todo-heading
Disabling this key binding for now because it overrides the standard
M-S-return binding. That was intentional but I am finding that this is
particularly problematic when inserting a new checklist item ('- [ ]'),
in which case, if it is under a level-2 heading in my notes, that
heading gets the PR-1 state. Also, I don't think I was finding this
function too useful anymore now that I mainly use
my-org-insert-todo-heading-respect-content.