Org: Only archive to sibling if in done / canceled state
This is especially for the sake of the 'a' org speed command: I don't
want to accidentally archive things that aren't ready to be archived
yet.

Motivation: I wanted to have a quick and easy way to archive
things (hence C-c a and the 'a' speed command), but I was worried that I
would then accidentally archive things that shouldn't be archived. This
should make it safer.