Org: make tags still distinct but stand out less
Motivation: I didn't like the look of the box around the ARCHIVE tag. At
first I tried using org-tag-faces to remove the box attribute for the
ARCHIVE tag, but that looked weird because there was still a box around
the ':' and the down arrow. So then I decided to completely change how
tags look to make them less in your face. Note that the new face matches
the face for org statistics (e.g., "[1/3]").

I also removed the customization for the "large" tag because now I use
an underline for all tags, and I'm not actually using the "large" tag
much, if at all.