Improve capture experience from other applications
Previously, there was a problem of an emacs window coming to the front.
This change fixes that problem by minimizing the capture window rather
than deleting it - since deleting it is what causes another window to
become active, and this doesn't seem to be an issue when
minimizing. (Though it seems we need to do the minimization via the
AppleScript after switching back to the previous application.)

For now, at least, I'm keeping the frame deletion behavior when
activated within emacs itself, because I prefer that behavior (both
because it avoids having this extra minimized frame, and it avoids the
flash of text clearing that I see when restoring the existing minimized
frame). But I may later decide that consistency is best and just do the
minimizing thing all the time.