Change org-id-link-to-org-use-id to create-if-interactive
With the setting of t, if I ran org-capture while in an org file, I
would get an id property in the current entry.
https://emacs.stackexchange.com/questions/30877/how-to-use-org-capture-without-creating-id-property
and
https://www.reddit.com/r/emacs/comments/16tbmni/prevent_orgmode_from_adding_an_id_property_to_the/
suggest this change, which seems to solve the problem.

(An alternative I considered before finding this solution is to let-bind
org-id-link-to-org-use-id to nil around the call to org-capture,
similarly to what I do in my-org-store-link-to-file.)