In my org capture finalization, use async-shell-command
I'm hopeful this will lead to less of an annoying delay compared with
shell-command; initial testing suggests this helps, but I'll need to try
it out more to see for sure.

Also removed and reworked some comments related to this delay.

If this doesn't help, here are some ideas that may help:
1. If the delay seems to be from the osascript, change the first part to `open -a APPLICATION`
2. Look into ways to improve the speed of shell-command in emacs
   - Check whether a simple shell-command also has delays; if so, that points to shell-command as a culprit. In this case, I could look into whether there's a way to speed up shell-command; e.g., does it try to load a shell environment, and if so, can we avoid that?
3. Can we use call-process or start-process instead of async-shell-command? If so, do those help?